Recover terrain and collision control flow and debris code generation
