+++ bazel/android_per_file_copts.bazelrc
@@ -299,3 +299,8 @@ build:target --per_file_copt=^src/legoapi/characters/core/charconfig\.cpp$@-O2
 build:target --per_file_copt=^src/legoapi/characters/motion/move\.cpp$@-O2
 build:target --per_file_copt=^src/MechInputTouch/MechInputTouchContextTasks\.cpp$@-O2
 build:target --per_file_copt=^src/gamelib/util/TouchHacks\.cpp$@-O2
+
+# Particle engine: original DebrisSetup2 (0x34be60, 2413 bytes) inlines matrix
+# rotation and unrolls setup loops; its game_deb initializer also matches -O3.
+# The unoptimized game API wrappers remain in game_debris_api.cpp.
+build:target --per_file_copt=^src/legoapi/render/fx/game_deb\.cpp$@-O3
